Split util/index.d.ts into multiple files
